requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.3.0
pytesseract==0.3.13
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
